워크북 구현

+++ src/main/java/com/example/umc9th/domain/review/dto/ReviewResponse.java
@@ -13,11 +13,11 @@ public class ReviewResponse {
     private Long reviewId;
     private Long storeId;
     private Long memberId;
-    private Integer score;
+    private Float score;
     private String content;
     private LocalDateTime createdAt;
 
-    public ReviewResponse(Long reviewId, Long storeId, String content, Integer rating) {
+    public ReviewResponse(Long reviewId, Long storeId, String content, Float rating) {
         this.reviewId = reviewId;
         this.storeId = storeId;
         this.content = content;
